Move Demo definition and implement a new CurrentDemo one.
CurrentDemo is an augmented Demo item with metadata on current slides,
ID and such. It's a helper to track current demo, and can advance, tick
or select a demo.
It's also handling selecting a particular slide.
It's always doing some Releasing teardown.

Add a helper as well to send current demo and convert it to a demo channel.